+++ python/soma/signature/attributetypes/integer.py
@@ -30,6 +30,7 @@
 # The fact that you are presently reading this means that you have had
 # knowledge of the CeCILL-B license and that you accept its terms.
 
+from __future__ import absolute_import
 import types
 import sys
 from soma.signature.attributetypes.number import Number
@@ -49,11 +50,11 @@ class Integer(Number):
         if minimum is None:
             self.minimum = None
         else:
-            self.minimum = long(minimum)
+            self.minimum = int(minimum)
         if maximum is None:
             self.maximum = None
         else:
-            self.maximum = long(maximum)
+            self.maximum = int(maximum)
 
     def checkValue(self, value):
         if not isinstance(value, float):
@@ -64,7 +65,7 @@ class Integer(Number):
         try:
             value = int(value)
         except ValueError:
-            value = long(value)
+            value = int(value)
         return value
